Add files via upload
This project analyzes the Netflix Titles Dataset using R to uncover trends in content distribution. The analysis includes data cleaning, transformation, and visualization to explore the types of content on Netflix, when and where it was added, popular genres, and the most frequent directors. It highlights key insights such as the growing number of titles over the years, dominance of movies, and the United States as the top content producer. This beginner-friendly project demonstrates essential data science skills using R.